requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
black==23.12.0
click==8.1.7
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
frozenlist==1.4.0
Markdown==3.5.1
multidict==6.0.4
pathspec==0.12.1
pytz==2023.3.post1
sqlparse==0.4.4
yarl==1.9.4
